Fix #61, implemented the ELEVATIONLIMITS keyword
The `configurttion.txt` template now mentions `minElevation` and
`maxElevation` keywords that are finally combined into `ELEVATIONLIMITS`
when the `basie` command is executed with a valid `configuration.txt`
file. The documentation of this repository is automatically generated,
the user manual is a bit updated, maybe we should think about updating
it with all the new features (this one as well as the nodding one, etc).